moved story data to storyUtils
