+++ src/Table/types.ts
@@ -14,7 +14,7 @@ export type TableStylesProps = {
 
 export type Primitive = string | number | boolean | bigint
 
-type RowAction<T> = {
+export type RowAction<T> = {
   //NOTE: what can i rename this too?
   onClick: (rowData: T) => void
   label?: ReactElement
